# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mirai Knowledge Systems API
  description: |
    建設土木ナレッジシステムのREST API

    ## 認証
    すべてのAPIエンドポイントはJWT認証が必要です。AuthorizationヘッダーにBearerトークンを含めてください。

    ## レート制限
    - 認証済みリクエスト: 1000/分、10000/時、100000/日
    - ファイルアップロード: 10/分
    - Microsoft 365連携: 50/分

    ## エラーレスポンス
    ```json
    {
      "success": false,
      "message": "エラーメッセージ",
      "error_code": "ERROR_CODE"
    }
    ```
  version: "2.0.0"
  contact:
    name: Mirai Knowledge Systems Support
    email: support@mirai-knowledge.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:5100
    description: 開発環境
  - url: https://api.mirai-knowledge.com
    description: 本番環境
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "エラーメッセージ"
        error_code:
          type: string
          example: "VALIDATION_ERROR"
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作が成功しました"
...
    Regulation:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "建設業法第19条施工体制台帳作成要領"
        content:
          type: string
          example: "施工体制台帳の作成要領と記載事項"
        category:
          type: string
          example: "法令"
        effective_date:
          type: string
          format: date
          example: "2025-01-01"
        status:
          type: string
          enum: [有効, 失効, 改正予定]
          example: "有効"
...
paths:
...
  /api/v1/knowledge:
    get:
      summary: ナレッジ一覧取得
      description: ナレッジの一覧を取得します。フィルタリングとページングに対応しています。
      tags:
        - Knowledge Management
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [施工計画, 品質管理, 安全衛生, 環境対策, 原価管理, 出来形管理, その他]
          description: カテゴリでフィルタリング
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, approved, archived]
          description: ステータスでフィルタリング
        - name: search
          in: query
          schema:
            type: string
          description: 検索キーワード
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: 取得件数
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: オフセット
      responses:
        '200':
          description: ナレッジ一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 1250
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Knowledge'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: ナレッジ作成
      description: 新しいナレッジを作成します
      tags:
        - Knowledge Management
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeCreate'
      responses:
        '201':
          description: ナレッジ作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Knowledge'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/knowledge/{knowledge_id}:
    get:
      summary: ナレッジ詳細取得
      description: 指定したIDのナレッジの詳細情報を取得します
      tags:
        - Knowledge Management
      security:
        - bearerAuth: []
      parameters:
        - name: knowledge_id
          in: path
          required: true
          schema:
            type: integer
          description: ナレッジID
      responses:
        '200':
          description: ナレッジ詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Knowledge'
        '404':
          description: ナレッジが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: ナレッジ更新
      description: 指定したIDのナレッジを更新します
      tags:
        - Knowledge Management
      security:
        - bearerAuth: []
      parameters:
        - name: knowledge_id
          in: path
          required: true
          schema:
            type: integer
          description: ナレッジID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeUpdate'
      responses:
        '200':
          description: ナレッジ更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Knowledge'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: ナレッジが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: ナレッジ削除
      description: 指定したIDのナレッジを削除します
      tags:
        - Knowledge Management
      security:
        - bearerAuth: []
      parameters:
        - name: knowledge_id
          in: path
          required: true
          schema:
            type: integer
          description: ナレッジID
      responses:
        '200':
          description: ナレッジ削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: ナレッジが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/incidents:
    get:
      summary: 事故レポート一覧取得
      description: 事故レポートの一覧を取得します
      tags:
        - Incident Management
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [調査中, 原因特定, 対策実施中, 完了]
          description: ステータスでフィルタリング
        - name: severity
          in: query
          schema:
            type: string
            enum: [軽微, 中程度, 重大, 致命的]
          description: 重大度でフィルタリング
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: 取得件数
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: オフセット
      responses:
        '200':
          description: 事故レポート一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 23
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/incidents/{incident_id}:
    get:
      summary: 事故レポート詳細取得
      description: 指定したIDの事故レポートの詳細情報を取得します
      tags:
        - Incident Management
      security:
        - bearerAuth: []
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: integer
          description: 事故レポートID
      responses:
        '200':
          description: 事故レポート詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '404':
          description: 事故レポートが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/consultations:
    get:
      summary: 専門家相談一覧取得
      description: 専門家相談の一覧を取得します
      tags:
        - Expert Consultation
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, assigned, in_progress, resolved, closed]
          description: ステータスでフィルタリング
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: 取得件数
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: オフセット
      responses:
        '200':
          description: 専門家相談一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 89
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Consultation'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/consultations/{consultation_id}:
    get:
      summary: 専門家相談詳細取得
      description: 指定したIDの専門家相談の詳細情報を取得します
      tags:
        - Expert Consultation
      security:
        - bearerAuth: []
      parameters:
        - name: consultation_id
          in: path
          required: true
          schema:
            type: integer
          description: 相談ID
      responses:
        '200':
          description: 専門家相談詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consultation'
        '404':
          description: 相談が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/sop/{sop_id}:
    get:
      summary: SOP詳細取得
      description: 指定したIDのSOPの詳細情報を取得します
      tags:
        - SOP Management
      security:
        - bearerAuth: []
      parameters:
        - name: sop_id
          in: path
          required: true
          schema:
            type: integer
          description: SOP ID
      responses:
        '200':
          description: SOP詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SOP'
        '404':
          description: SOPが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/regulations/{reg_id}:
    get:
      summary: 法令詳細取得
      description: 指定したIDの法令の詳細情報を取得します
      tags:
        - Regulation Management
      security:
        - bearerAuth: []
      parameters:
        - name: reg_id
          in: path
          required: true
          schema:
            type: integer
          description: 法令ID
      responses:
        '200':
          description: 法令詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Regulation'
        '404':
          description: 法令が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...